% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{add_observed_values}
\alias{add_observed_values}
\title{Add observed values for phyloscanner results}
\usage{
add_observed_values(df1)
}
\arguments{
\item{df1}{}
}
\value{
dataframe with the added observed values
}
\description{
Add observed values for phyloscanner results
}
\details{
If observed value is TP,(true positives), it means that
   phyloscanner correctly identified the transmission pair. If observed value
   is FP (false positives), phyloscanner incorrectly identify the transmission
   pair. If observed value is FN (false negative) it means it is a true
   transmission pair, but phyloscanner was not able to identify it. If
   observed value is TN (true negative), it means it is not a transmission
   pair and phylosvanner correctly identified as not a transmission pair.
}
