% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{add_observed_values_directness}
\alias{add_observed_values_directness}
\title{Add observed values for phyloscanner results for directness}
\usage{
add_observed_values_directness(df1)
}
\arguments{
\item{df1}{}
}
\value{
dataframe with the added observed values
}
\description{
Add observed values for phyloscanner results for directness
}
\details{
If observed value is TP (true positives), it means that
   phyloscanner correctly identified a transmission pair with correct direction
   of transmission(A infected B or B infected A). If observed value
   is FP (false positives), phyloscanner incorrectly identify a transmission
   pair. If observed value is FN (false negative), it means it is a true
   transmission pair (with correct direction of transmission from A to B or
   B to A), but phyloscanner was not able to identify it. If
   observed value is TN (true negative), it means it is not a transmission
   pair and phyloscanner correctly identified it as not a transmission pair
   (with correct direction of transmission from A to B or B to A).
}
