% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{add_root_edge2}
\alias{add_root_edge2}
\title{Add root edge to phylogenetic tree}
\usage{
add_root_edge2(tree, root.edge_value = 0)
}
\arguments{
\item{tree}{object of class Phylo}

\item{root.edge_value}{scalar for the value to use as root.edge in the
phylogenetic tree}
}
\description{
Add root edge to phylogenetic tree
}
