% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{add_sampled_times}
\alias{add_sampled_times}
\title{Function to add the sampled times for transmission pairs}
\usage{
add_sampled_times(df1, sampled_times)
}
\arguments{
\item{df1}{dataframe containing information of transmission pairs.}

\item{sampled_times}{dataframe containing information of sampled times
for IDs in df1.}
}
\value{
dataframe with added information of the sampled times for donor
   (infected individual) and recipient (susceptible individual).
}
\description{
Function to add the sampled times for transmission pairs
}
\details{
This function will adataframe composed of a single row and check
   whether it represents a transmission pair (independent of A transmitted
   to B or B transmitted to A). If it is a transmision pair will add to the
   dataframe the sampled times of infected and susceptible individuals.
   If not a transmission pair, than the value of NA will be added to the
   dataframe.
}
