% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{check_linked_transmissions}
\alias{check_linked_transmissions}
\title{Check whether a pair of individuals are linked by two or more individuals.}
\usage{
check_linked_transmissions(df1, tm)
}
\arguments{
\item{df1}{Dataframe containing information of summarized results obtained
with phyloscanner after running the function summarize_trans.}

\item{tm}{Dataframe of transmission matrix containing information of who
transmitted to whom and at what time transmission happened.}
}
\value{
Dataframe with a new column "trans" with only the pair of individuals
   that are linked.
   The value for column "trans" will be "linked" or "not linked".
}
\description{
This function will check whether a pair of individuals (host.1 and host.2)
are linked because intermediary individuals were involved. For example, host.3
infected host.1 and host.2.
}
