% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{check_swap_transmissions}
\alias{check_swap_transmissions}
\title{Check whether a pair of individuals are linked and represent a direct
transmission.}
\usage{
check_swap_transmissions(df1, df2)
}
\arguments{
\item{df1}{Dataframe containing information of summarized results obtained
with phyloscanner after running the function summarize_trans.}

\item{df2}{Dataframe containing information of the true transmissions.}
}
\value{
Dataframe with a new column "trans" with only the individuals that
   represent a "swap" transmission.
   The value for column "trans" will be "swap" (for transmission from
   host.2 to host.1).
}
\description{
This function will check whether a pair of individuals (host.1 and host.2)
represent a true transmission from host.2 to host.1 independent of phyloscanner
results.
}
