% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{check_true_pair_directness}
\alias{check_true_pair_directness}
\title{Check whether a pair of individuals represent a transmission pair.}
\usage{
check_true_pair_directness(df1, tm)
}
\arguments{
\item{df1}{Dataframe containing information of summarized results obtained
with phyloscanner after running the function summarize_trans.}

\item{tm}{Dataframe containing information of for the transmission matrix.}
}
\value{
Dataframe with a new column "real_pair" that take the value "true"
   (for a true transmission pair independent of directness,
   which means host.1 transmitted to host.2 or host.2 transmitted to host.1).
   It will also have another column "time_trans" for the time of transmission
   of a real pair. If a pair is not a real transmission pair, than the value
   for time_trans will be "NA".
}
\description{
This function will check whether a pair of individuals (host.1 and host.2)
represent a transmission pair from host.1 to host.2 or host.2 to host.1
independent of phyloscanner results.
}
