% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{check_true_transmissions}
\alias{check_true_transmissions}
\title{Check whether a pair of individuals represent a true transmission.}
\usage{
check_true_transmissions(df1, tm)
}
\arguments{
\item{df1}{Dataframe containing information of summarized results obtained
with phyloscanner after running the function summarize_trans.}

\item{tm}{Dataframe containing information of the transmission matrix (tm).
The tm have information of who transmitted to whom and what time that
happened.}
}
\value{
Dataframe with a new column "trans" with only the individuals that
   represent a true transmission with correct direction of transmission.
   The value for column "trans" will be "true" (for true transmission).
}
\description{
This function will check whether a pair of individuals (host.1 and host.2)
represent a true transmission from host.1 to host.2 independent of phyloscanner
results.
}
