% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{compute_log_importance_weight}
\alias{compute_log_importance_weight}
\title{Calculate the log importance weight based on observed and simulated data}
\usage{
compute_log_importance_weight(diag_obs, diag_sim)
}
\arguments{
\item{diag_obs}{Frequency of observed incidence of HIV diagnosis in MSM
from 1982 to 2020}

\item{diag_sim}{Frequency of estimated incidence of HIV diagnosis in MSM
from 1982 to 2020}
}
\value{

}
\description{
Calculate the log importance weight based on observed and simulated data
}
