% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{compute_log_importance_weight_incidence}
\alias{compute_log_importance_weight_incidence}
\title{Calculate the log importance weight based on observed and simulated data
for incidence}
\usage{
compute_log_importance_weight_incidence(incid_obs, incid_sim)
}
\arguments{
\item{diag_obs}{Frequency of observed incidence in MSM from 1980 to 2020}

\item{diag_sim}{Frequency of estimated incidence in MSM from 1980 to 2020}
}
\value{
Scalar for the log importance weight
}
\description{
Calculate the log importance weight based on observed and simulated data
for incidence
}
