% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_times_tsv}
\alias{create_times_tsv}
\title{Get tsv for times}
\usage{
create_times_tsv(ids, time_seq, seq_count, prefix = NULL)
}
\arguments{
\item{ids}{Vector of samples IDs.}

\item{time_seq}{Vector for sample time. If one value is provided,
it will be replicated to the size of samples in the transmission matrix (tm).}

\item{seq_count}{The number of sequences per sample in the transmission matrix.
If one value is provided, it will be replicated to the size of samples in
the transmission matrix (tm).}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{
Write data to tsv file.
}
\description{
In this function, the times file for CoaTrans program
will contain only the sampled individuals (to mimic a real HIV study)
}
\details{
If a prefix is not provided, tab delimited file file will be saved as times.tsv
}
