% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_transmissions_tsv}
\alias{create_transmissions_tsv}
\title{Create transmission matrix tsv}
\usage{
create_transmissions_tsv(tm, time_tr, prefix = NULL)
}
\arguments{
\item{tm}{Transmission matrix as returned using the function
\link[EpiModel]{get_transmat}}

\item{time_tr}{Time of transmission for seeds. The other time of transmission
should be provided in the tm dataframe.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{
Write data to tsv file.
}
\description{
Create a transmission matrix file to be used with the
   CoaTrans. CoaTrans. requires that seeds are included
   in the file.
}
\details{
If a prefix is not provided, tsv file will be saved as transmissions.tsv
}
