% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{days2years}
\alias{days2years}
\title{Convert days to years}
\usage{
days2years(sampleTimes, init_date)
}
\arguments{
\item{sampleTimes}{Vector for sample times in days}

\item{init_date}{The date in the format of year-month-day to convert days to
years}
}
\value{
Vector of sample times in decimal years
}
\description{
This function converts days to years taking into consideration the start date
of simulations.
}
\examples{
sample_times <- c(8760, 10526, 14600)
init_date <- "1981-01-01"
days2years(sample_times, init_date)
}
