% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model0_net.R
\name{generate_model0_net}
\alias{generate_model0_net}
\title{Generate a simple HIV epidemic model for use by phydynR}
\usage{
generate_model0_net(
  natmort = 0.0131061,
  srcGrowth = 0.018,
  src1980 = 50000,
  i1980 = 1,
  gamma = 1/5.06,
  importRate = 1/10,
  art_start = 2004
)
}
\value{
A list with elements
* dm : a demographic process model
* parms: a default list of parameters
* x0 : a default vector of start conditions
}
\description{
* src : global reservoir
* i : infected and infectious
* z : infected and on ART
Transmission occurs at a rate beta(t) i(t) where beta is a linear
interpolating spline. The treatment rate i->z is initially zero but
increases linearly to 2015
}
