% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cd4s_sampling}
\alias{get_cd4s_sampling}
\title{Get CD4s by node IDs at time of sampling}
\usage{
get_cd4s_sampling(sampled_info, stages)
}
\arguments{
\item{sampled_info}{dataframe of sampled IDs and sampled times}

\item{stages}{dataframe that tracks the stage of HIV infection for each ID.}
}
\value{

}
\description{
Get CD4s by node IDs at time of sampling
}
\details{
This function will get the stage of HIV infection at time of sampling.
}
