% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{get_difference}
\alias{get_difference}
\title{Estimate difference between sampling time and time of infection}
\usage{
get_difference(df1, df2)
}
\arguments{
\item{df1}{Dataframe with information on time of sampling}

\item{df2}{Transmission matrix to get information on when the individual
got infected}
}
\value{
Dataframe with difference in months.
}
\description{
Estimate the difference in months between time of infection of a susceptible
individual and time of sampling of that same individual. It assumes that a
month has 30 days.
}
