% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{get_id_by_location}
\alias{get_id_by_location}
\title{Get ID of individuals by location}
\usage{
get_id_by_location(tm, location)
}
\arguments{
\item{tm}{Transmission matrix of who transmitted to whom.}

\item{location}{Character of location at time of transmission to get IDs
to sample from. Location can take the value of "region" or "global".}
}
\value{
Vector of IDs from location of choice.
}
\description{
This function will get the ID of infected individuals from specific location
(region or global) at the time of seroconversion.
}
\details{
Note that location will be those at the time of seroconvertion.
}
