% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{get_n_by_origin}
\alias{get_n_by_origin}
\title{Get total number of individuals in network by region}
\usage{
get_n_by_origin(tm, location)
}
\arguments{
\item{tm}{Transmission matrix of who transmitted to whom.}

\item{location}{Character of location at time of transmission to get IDs
to sample from. Location can take the value of "region" or "global".}
}
\value{
Integer of the total number of individuals in location of choice.
   Note that the function will return the total number of individual at location
   of choice at time of seroconversion.
}
\description{
Get the total number of individuals in the final network by region of choice.
}
\details{
Note that location will be those at the time of seroconvertion.
}
