% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{get_newids}
\alias{get_newids}
\title{Get a sampled ID (sid) and vector of IDs to sample from}
\usage{
get_newids(ids)
}
\arguments{
\item{ids}{Vector of ids to sample from}
}
\value{
One element for sampled ID (ID).
}
\description{
Get a sampled ID (sid) and vector of IDs to sample from
}
\details{
This function will get a vector of IDs to sample from and randomly
   sample an ID. If there is only one element in ids, then sid will be equal
   to ids.
}
\seealso{
\code{\link[base]{sample}} for more information on random samples.
}
