% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{get_origin_at_samplingTime}
\alias{get_origin_at_samplingTime}
\title{Get origin at time of sampling}
\usage{
get_origin_at_samplingTime(tm, origin, sid, st)
}
\arguments{
\item{tm}{Transmission matrix of who transmitted to whom.}

\item{origin}{Dataframe of IDs that migrated during the simulation and time
time that migration happened. Time must be in decimal year.}

\item{sid}{sampled ID}

\item{st}{sampled time}
}
\value{

}
\description{
Get the origin code at time of sampling.
}
