% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{get_subset_tips}
\alias{get_subset_tips}
\title{Get subset of susceptibles and infected individuals}
\usage{
get_subset_tips(tips, tm)
}
\arguments{
\item{tips}{Vector of tip names (without the "ID")}

\item{tm}{Dataframe of transmission matrix containing information of who
transmitted to whom and at what time transmission happened.}
}
\value{
List of pairs of sus and inf individuals
}
\description{
This function get the pairs of inf and sus individuals given a transmission
matrix.
}
