% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{get_tipNames}
\alias{get_tipNames}
\title{Get tip names of phylogenetic tree based on transmission matrix}
\usage{
get_tipNames(x, format = "migrant", by_areas = "region", max_value = NULL)
}
\arguments{
\item{x}{An object of class \code{"transmat"}, the output from
\code{\link{get_transmat}}.}

\item{format}{If format = "origin" return tip in the form of ID_global or ID_region.
If format = "migrant" return tip in the form of ID_1, ID_2, ID_21, ID_12}

\item{by_areas}{by_areas = "all" if interested in all seeds, or by_areas = "region"
if interested in seeds for region only.}

\item{max_value}{max value for comparison to select the seed values. It will
select only seed values less than max_value for seeds from region only.}
}
\value{
List of tip names. See details for codes on the tip names related
   to origin or global.
}
\description{
Get tip names of phylogenetic tree based on transmission matrix
}
\details{
This code was build using the algorithm as in \code{\link{as.phylo.transmat}}.
The difference is that it will return only the tip names of phylogenetic tree.
If using format = "origin", tip names will be in the form of ID_global or
ID_region. If using format = "migrant", tip names will be in the form of
ID_1, ID_21, ID_2, ID_12.
Migrant values can take the value of 1 if vertex is from region,
2 if vertex is from global,
21 if vertex is from global that migrated to region and
12 if vertex is from region that migrated to global.
It will return the origin at time of seroconversion.


This script will specifically work with the package HIVepisim.
}
