% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{order_IDs_byDate}
\alias{order_IDs_byDate}
\title{Remove duplicated rows and order by date of transmission}
\usage{
order_IDs_byDate(tm_subset_list)
}
\arguments{
\item{tm_subset_list}{List of dataframe of subset of transmission matrix}
}
\value{
Dataframe in which duplicated rows are removed and ordered by
   transmission date
}
\description{
Remove duplicated rows and order by date of transmission
}
