% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recency_test}
\alias{recency_test}
\title{Get which IDs are at early stage of HIV infection in the past 6 months}
\usage{
recency_test(sampled_info, stages)
}
\arguments{
\item{sampled_info}{dataframe of sampled IDs and sampled times}

\item{stages}{dataframe that tracks the stage of HIV infection for each ID.}
}
\value{
Vector of IDs.
}
\description{
This function mimics the idea of the recency test. It selects the IDs of
individuals with recent HIV infection (in the past 6 months)
}
