% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{selectIDs}
\alias{selectIDs}
\title{Get IDs before sampled time.}
\usage{
selectIDs(IDs, metadata, st)
}
\arguments{
\item{IDs}{Vector of IDs}

\item{metadata}{Dataframe of specific metadata to get IDs before sampled time.}

\item{st}{Sampled time in decimal year.}
}
\value{
Dataframe of IDs and the time of interest in decimal year.
}
\description{
Get IDs before sampled time.
}
\details{
This function will get a vector of IDs and get the IDs that
   fit the condition of existing before the sampled time.
}
