% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{set_sampledIDs}
\alias{set_sampledIDs}
\title{Set a dataframe of sampled IDs (sid) and their sampled times.}
\usage{
set_sampledIDs(sid, st)
}
\arguments{
\item{sid}{The sampled ID}

\item{st}{Sampled time in decimal year.}
}
\value{
Return dataframe of sampled IDs and their sampled times.
}
\description{
Set a dataframe of sampled IDs (sid) and their sampled times.
}
