% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{split_dirs}
\alias{split_dirs}
\title{Function to split vector of with names of directories into chunks of
specific size}
\usage{
split_dirs(dirs, size)
}
\arguments{
\item{dirs}{Vector with names of directories}

\item{size}{Size to divide vector.}
}
\value{

}
\description{
Function to split vector of with names of directories into chunks of
specific size
}
\details{
If we want to split a vector of 15 elements into chunks of 10 elements,
   the first object in the list will have 10 element, and the second object
   in the list will have 5 elements.
}
