% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{summarize_all_data}
\alias{summarize_all_data}
\title{Summarize all results obtained with phyloscanner}
\usage{
summarize_all_data(results_phylo)
}
\arguments{
\item{results_phylo}{Dataframe containing the values for number of transmission
pairs analysed, number of true pairs within the total pairs analysed, number
of true transmission pairs identified by phyloscanner, etc. by replicate,
combination of parameter values and migration rate.}
}
\value{
Tibble object with the total number of pairs analysed, total number of
   true transmissions, total number of true positives as identified by phyloscanner,
   total number of false negatives by phyloscanner, total number of swaps
   (host.2 infected host.1, instead of host.1 infected host.2), and total
   number of false positives as identified by phyloscanner.
}
\description{
Summarize all results obtained with phyloscanner
}
\details{
This function is used in the R script summarize_phyloscanner_results_v2
   (check scripts in \href{https://github.com/thednainus/HIVepisimAnalysis/blob/main/Analyses/deep_sequencing/Result_analyses/summarize_phyloscanner_results_v2.R}{HIVepisimAnalysis GitHub})
}
