% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{summarize_trans_compl}
\alias{summarize_trans_compl}
\title{Summarize results obtained with phyloscanner.}
\usage{
summarize_trans_compl(data)
}
\arguments{
\item{data}{}
}
\value{
Dataframe containing information of whether host.1 and host.2
   are linked and whether there direction of transmission is also correct.
}
\description{
Summarize results obtained with phyloscanner.
}
\details{
Here we summarized the results obtained with phyloscanner.
   The summary is based on the summary results that phyloscanner returns
   in which it summarizes the "ancestry" type by window.
   We considered that two individuals are linked if proportion of values for
   ancestry is complex, multiTrans and trans are >= 0.50
   On the other hand, we considered that two individuals represent a true
   direction of transmission from host.1 to host.2 if the proportion of
   multiTrans, trans, and complex are >= 0.375.
   Here, we are interested to understand whether including complex transmissions
   would make it better to detect direction of transmission.
   Note that to carry out this summary, we pre-removed all
   ancestry = noAncestry.
}
