% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{time_atSeroconversion}
\alias{time_atSeroconversion}
\title{Get the time of duplicates at seroconversion.}
\usage{
time_atSeroconversion(dup_rows, get_time)
}
\arguments{
\item{dup_rows}{matrix of list of matrices of transmat
that have duplicated vertex IDs.}

\item{get_time}{vector of oldest time}
}
\description{
This is a support function for the main function
   \code{\link{toPhylo_transmatOrigin}}.
}
\details{
During the simulations, vertex can migrate from region to global
or global to region. When adding the information of migrant or origin to the
tip name, we will observe duplicated on vertex IDs because it can be in region
at one time point but later in global, for example.
This function will get the duplicated rows in the transmission matrix
in which a duplicated vertex is observed. It will then get the time observed
for each duplicated vertex based on get_time. Basically it will return the
origin at time of seroconversion.
}
