% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{est_sampleSize}
\alias{est_sampleSize}
\title{Estimation sample size based on percentage}
\usage{
est_sampleSize(
  perc,
  start_date,
  end_date,
  art_init,
  departure,
  diag_info,
  tm,
  location = "region",
  onART = FALSE
)
}
\arguments{
\item{perc}{Percentage of the population to be sampled: Numbers should be between
0 and 1.}

\item{start_date}{Start date for sampling times in decimal year.}

\item{end_date}{End date for sampling times in decimal year.}

\item{art_init}{Dataframe of IDs and time each ID initiated antiretroviral
treatment (ART). Time must be in decimal year.}

\item{departure}{Dataframe of IDs and their time of departure. Time must be
in decimal year.}

\item{diag_info}{Dataframe of IDs and their time of diagnosis. Time must be
in decimal year.}

\item{tm}{Transmission matrix of who transmitted to whom.}

\item{location}{Character of location at time of transmission to get IDs
to sample from. Location can take the value of "region" or "global".}
}
\value{

}
\description{
Estimation sample size based on percentage
}
\details{
This function will estimate the sample size based on a percentage
(0 to 1) and the end of the simulation (end_date). To carry out our analyses,
we used a sampled to sample time of sequencing just to mimic real HIV studies.
However, we can only sample diagnosed and active (and on ART) inidviduals,
it is hard to estimate the exact number of individuals based on the sampling
times that will be randomly sampled.
}
