% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{set_sampledIDs}
\alias{set_sampledIDs}
\title{Set a dataframe of sampled IDs (sid) and their sampled times.}
\usage{
set_sampledIDs(sid, st, migrant_code)
}
\arguments{
\item{sid}{The sampled ID}

\item{st}{Sampled time in decimal year.}

\item{migrant_code}{Origin code, If 1 is "region" but never migrated,
if 12 node ID migrated from "region" to "global",
if 2 is "global" and never migrated,
if 21 node ID migrated from "global" to "region".}
}
\value{
Return dataframe of sampled IDs and their sampled times.
}
\description{
Set a dataframe of sampled IDs (sid) and their sampled times.
}
